Use Docker multistage build
This reduces the image size from 280MB to 20MB, the container
startup time is about a second compare to 14 seconds, the running
container uses 2MB instead of 55MB and uses 10x less IO to startup.

Measurements done on Fedora 28, using btrfs storage backend for Docker
CE 18.06.0, on a laptop with a SSD and dual core CPU (i7-5600U).